Moved package to new 'skipstage'
