+++ pkg/app/piped/executor/analysis/analysis.go
@@ -29,6 +29,7 @@ import (
 	"github.com/pipe-cd/pipecd/pkg/app/piped/analysisprovider/metrics"
 	metricsfactory "github.com/pipe-cd/pipecd/pkg/app/piped/analysisprovider/metrics/factory"
 	"github.com/pipe-cd/pipecd/pkg/app/piped/executor"
+	"github.com/pipe-cd/pipecd/pkg/app/piped/executor/skipstage"
 	"github.com/pipe-cd/pipecd/pkg/config"
 	"github.com/pipe-cd/pipecd/pkg/model"
 )
@@ -63,7 +64,7 @@ func Register(r registerer) {
 // Any of those fail then the stage ends with failure.
 func (e *Executor) Execute(sig executor.StopSignal) model.StageStatus {
 	// Skip the stage if needed based on the skip config.
-	skip, err := executor.CheckSkipStage(sig.Context(), e.Input, e.StageConfig.AnalysisStageOptions.SkipOptions)
+	skip, err := skipstage.CheckSkipStage(sig.Context(), e.Input, e.StageConfig.AnalysisStageOptions.SkipOptions)
 	if err != nil {
 		e.Logger.Error("failed to check whether skipping the stage", zap.Error(err))
 		return model.StageStatus_STAGE_FAILURE
