+++ pkg/app/piped/executor/skipstage/skipstage_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package executor
+package skipstage
 
 import (
 	"testing"
