+++ pkg/app/piped/executor/waitapproval/waitapproval.go
@@ -24,6 +24,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/pipe-cd/pipecd/pkg/app/piped/executor"
+	"github.com/pipe-cd/pipecd/pkg/app/piped/executor/skipstage"
 	"github.com/pipe-cd/pipecd/pkg/config"
 	"github.com/pipe-cd/pipecd/pkg/model"
 )
@@ -53,7 +54,7 @@ func Register(r registerer) {
 // Execute starts waiting until an approval from one of the specified users.
 func (e *Executor) Execute(sig executor.StopSignal) model.StageStatus {
 	// Skip the stage if needed based on the skip config.
-	skip, err := executor.CheckSkipStage(sig.Context(), e.Input, e.StageConfig.WaitApprovalStageOptions.SkipOptions)
+	skip, err := skipstage.CheckSkipStage(sig.Context(), e.Input, e.StageConfig.WaitApprovalStageOptions.SkipOptions)
 	if err != nil {
 		e.Logger.Error("failed to check whether skipping the stage", zap.Error(err))
 		return model.StageStatus_STAGE_FAILURE
